requirements.txt
asgiref==3.9.1
dj-database-url==3.0.1
Django==5.2.4
gunicorn==23.0.0
packaging==25.0
psycopg2-binary==2.9.10
python-dotenv==1.1.1
sqlparse==0.5.3
whitenoise
openpyxl==3.1.2
openai
